UI Grab Controller
Migrated the grabbing logic into the UI Grab controller prefab. This has a canvas and box collider with an XR Grab interactable script attached, allowing this UI to be grabbed. This prefab has been attached to the AR Display to make it grabbable. The main scene has been updated to include 3 different sizes of the AR Display for testing. The portfolio XR setup has had the poke depth increased. There is currently an issue with the direct grabbing and poking interactions causing direct grabbing stop the poking interactions. Looking into this.